factor out focus_scan from macros
